Batch size > 1, plots, adversarial loss update
The batch size can now be bigger than 1.
The adversarial loss calculation was updated. The MSEloss is now calculated between the discriminator (with a fake signal as input) output and zero. The discriminator should output 0 for a fake signal. Prior to the update, the MSEloss was calculated between the discriminator output and 1.
The losses are all plotted.
Config information is now part of the notebook and the config file will be removed in future updates